modified for checking if ppd-uri file exists in order to not download it all the times

+++ providers/printers.rb
@@ -106,20 +106,21 @@ action :setup do
 
         create_ppd_with_ppd_uri = false
 	if printer.attribute?('ppd_uri')
-		Chef::Log.info(" - using PPD_URI: #{printer.ppd_uri}")
-		FileUtils.mkdir_p('/usr/share/cups/model')
-
-		ppd_uri_dw = Mixlib::ShellOut.new("/usr/bin/wget -c --no-check-certificate -O /usr/share/cups/model/#{curr_ptr_name}.ppd #{printer.ppd_uri}")
-		ppd_uri_dw.run_command
-		if ppd_uri_dw.exitstatus == 0
-			file "/usr/share/cups/model/#{curr_ptr_name}.ppd" do
-				mode '0644'
-				owner 'root'
-				group 'root'
+		if not ::File.exists?("/usr/share/cups/model/#{curr_ptr_name}.ppd")
+			Chef::Log.info(" - using PPD_URI: #{printer.ppd_uri}")
+			FileUtils.mkdir_p('/usr/share/cups/model')
+			ppd_uri_dw = Mixlib::ShellOut.new("/usr/bin/wget -c --no-check-certificate -O /usr/share/cups/model/#{curr_ptr_name}.ppd #{printer.ppd_uri}")
+			ppd_uri_dw.run_command
+			if ppd_uri_dw.exitstatus == 0
+				file "/usr/share/cups/model/#{curr_ptr_name}.ppd" do
+					mode '0644'
+					owner 'root'
+					group 'root'
+				end
+				create_ppd_with_ppd_uri = true
+			else
+				Chef::Log.info(" - failed to obtain PPD using #{printer.ppd_uri}")
 			end
-			create_ppd_with_ppd_uri = true
-		else
-			Chef::Log.info(" - failed to obtain PPD using #{printer.ppd_uri}")
 		end
         else
 		ppd_uri = ''
